requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.1
drf-yasg2==1.19.3
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
zipp==3.3.1
